Adds map reader that loads map from file
Also add two simple maps for testing
Tile and resources configuration are now present in the conf/ folder.
At start, the game loads all the configurations in this folder,
so to add a resource, a conf file must be added, as well as a corresponding
image in res/image/ folder.